Revit 2025 + Add-in Manager fixes (#227)
* Update azure-pipelines to include variable group (requires reformatting)

* Revit 2025 (#223)

* Add Revit 2025 build
* Migrate to .NET 8 for 2025
* Add Nice3point.Revit dependencies
* Remove HOK.BetaToolsManager and HOK.SmartBCF

* Addin Manager updates

* Bump version

* Specify dotnet version

+++ _build/copy_artifacts.ps1
@@ -1,11 +1,12 @@
 ﻿$buildConfiguration = $Env:BUILD_CONFIGURATION
 
 $addinFolder = Join-Path .\_artifacts\ $buildConfiguration
-$libraryFolder = Join-Path $addinFolder \HOK-Addin.bundle\Contents\
 
-$list = Import-Csv .\_build\files.csv | ForEach { 
+Import-Csv .\_build\files.csv | ForEach-Object { 
     $s = "{0}{1}" -f $_.sourcePath,$buildConfiguration
     $p = Join-Path -Path $s -ChildPath $_.fileName 
     $d = "{0}\{1}" -f $addinFolder,$_.destination
-    Copy-Item $p -Destination $d
+    If ((Test-Path $p)) {
+        Copy-Item $p -Destination $d
+    }
 }
